Change TBA to session-based authentication
src/auth/auth.py
- replace jwt tokens with session tokens generated by nanoid.
Decoding the jwt token requires a lot of computing power,
and must be performed for all requests. Session tokens are generated
once and during requests we do a simple equality check which is a
lightweight operation. As a result, this will save us a lot
of computational power.

src/auth/auth_router.py
- replace jwt to session tokens.